feat(bin): get-tcp.sh
The /proc/net/tcp file provides detailed information about the current TCP
connections on the system, each line contains data about a TCP socket.
It works like this:

local_address: local IP and port of the connection.

rem_address: remote IP and port.

st: The state of the TCP connection (for example, 01 means ESTABLISHED).

tx_queue and rx_queue: The transmit and receive queues.

uid: The ID of the user who initiated the connection.

inode: The identifier of the inode node that holds the socket.